final code changes, hope it works on GH Actions.
