+++ test/src/services/edx-api-service.js
@@ -1,8 +1,20 @@
+import generator from 'generate-password';
+
 const restUtils = require('../helpers/rest-utils');
 const constants = require('../config/constants');
 import {getActivationCode} from '../helpers/secure-exchange-utils';
+import faker from 'faker';
+import {ClientFunction} from 'testcafe';
+import {credentials} from '../config/constants';
+import {getToken} from '../helpers/oauth-utils';
+import LoginPage from '../page_models/login-page';
+import UserActivation from '../page_models/user-activation';
+const loginPage = new LoginPage();
+const userActivationPage = new UserActivation();
+
 const log = require('npmlog');
 const EXCHANGE_ENDPOINT = `${constants.edx_api_base_url}api/v1/edx/exchange`;
+
 
 async function getAllEdxUserRoles(token) {
   const endpoint = 'api/v1/edx/users/roles';
@@ -94,7 +106,46 @@ const edxApiService = {
     const endpoint = 'api/v1/edx/users';
     const url = `${constants.edx_api_base_url}${endpoint}/${edxUser.edxUserID}`;
     await restUtils.deleteData(token, url);
+  },
+  async generateCode(){
+    return  generator.generateMultiple(2,{
+      length: faker.datatype.number({ 'min': 7, 'max': 7 }),
+      numbers: true,
+      uppercase: true,
+    });
+  },
+  async  login(t) {
+    await t.navigateTo(t.fixtureCtx.activationUrl[0]);
+    log.info('EDX Login page loaded successfully!');
+    // log in, assert return to baseurl
+    const getLocation = ClientFunction(() => document.location.href);
+    await t.typeText(loginPage.userNameInput(), credentials.activateUserCredentials.username, {timeout: 20000})
+      .typeText(loginPage.passwordInput(), credentials.activateUserCredentials.password, {timeout: 20000})
+      .click(loginPage.submitCredentialsButton());
+    log.info('User could login successfully!');
+    return getLocation;
+  },
+
+  async  createFixtureSetupForEdxUserActivation(ctx,primaryCode,personalCode) {
+    try {
+      const data = await getToken();
+      ctx.activationUrl = await edxApiService.createUserActivationUrl(data.access_token,primaryCode,personalCode);
+      ctx.acCode1 = ctx.activationUrl[1].edxActivationCodeId;
+      ctx.acCode2 = ctx.activationUrl[2].edxActivationCodeId;
+      ctx.primaryCode= primaryCode;
+      ctx.personalCode= personalCode;
+    } catch (e) {
+      console.error(e);
+    }
+  },
+
+  async  submitDetailsOnUserActivationForm(t, mincode, primaryActivationCode, personalActivationCode) {
+    await t.typeText(userActivationPage.mincodeInput(), mincode, {timeout: 20000})
+      .typeText(userActivationPage.primaryActivationCodeInput(), primaryActivationCode, {timeout: 20000})
+      .typeText(userActivationPage.personalActivationCodeInput(), personalActivationCode, {timeout: 20000})
+      .click(userActivationPage.submitUserActivationButton());
   }
+
 };
 
 module.exports = edxApiService;
